requirements.txt
fastapi
uvicorn
torch
numpy
Pillow
requests
scikit-image
matplotlib
google-auth
geemap
earthengine-api
python-multipart
requests
streamlit

